Generate workflow in build.sbt
Scala all the things \o/
