+++ .github/workflows/ci.yml
@@ -1,104 +1,55 @@
 ---
 {
-   "jobs": {
-      "Docusaurus": {
-         "runs-on": "ubuntu-latest",
-         "steps": [
-            {
-               "uses": "actions/checkout@v1"
-            },
-            {
-               "uses": "olafurpg/setup-scala@v2"
-            },
-            {
-               "name": "Website",
-               "run": "csbt docs/docusaurusCreateSite"
-            }
-         ]
-      },
-      "PresentationCompiler": {
-         "matrix": {
-            "scala": [
-               "2.13.0",
-               "2.13.1",
-               "2.12.8",
-               "2.12.9",
-               "2.12.10",
-               "2.12.7",
-               "2.11.12"
+    "name": "CI",
+    "on": [
+        "push",
+        "pull_request"
+    ],
+    "jobs": {
+        "Scalafmt": {
+            "runs-on": "ubuntu-latest",
+            "steps": [
+                {
+                    "uses": "actions/checkout@v1"
+                },
+                {
+                    "uses": "olafurpg/setup-scala@v2"
+                },
+                {
+                    "name": "Formatting",
+                    "run": "./bin/scalafmt --test"
+                }
             ]
-         },
-         "runs-on": "ubuntu-latest",
-         "steps": [
-            {
-               "uses": "actions/checkout@v1"
-            },
-            {
-               "uses": "olafurpg/setup-scala@v2"
-            },
-            {
-               "name": "Run tests",
-               "run": "csbt ++${{  matrix.scala }} cross/test"
-            }
-         ]
-      },
-      "Scalafix": {
-         "runs-on": "ubuntu-latest",
-         "steps": [
-            {
-               "uses": "actions/checkout@v1"
-            },
-            {
-               "uses": "olafurpg/setup-scala@v2"
-            },
-            {
-               "name": "Linting",
-               "run": "csbt scalafixCheck"
-            }
-         ]
-      },
-      "Scalafmt": {
-         "runs-on": "ubuntu-latest",
-         "steps": [
-            {
-               "uses": "actions/checkout@v1"
-            },
-            {
-               "uses": "olafurpg/setup-scala@v2"
-            },
-            {
-               "name": "Formatting",
-               "run": "./bin/scalafmt --test"
-            }
-         ]
-      },
-      "UnitTests": {
-         "matrix": {
-            "test": [
-               "{foldingrange,filewatch,hover,metals,references,statusbar,completions}",
-               "{ansi,workspacesymbol,bsp,builds,applyedit,compile,definition}",
-               "{http,formatting,documentsymbol,documenthighlight,codelens,tvp,mtags}"
+        },
+        "Scalafix": {
+            "runs-on": "ubuntu-latest",
+            "steps": [
+                {
+                    "uses": "actions/checkout@v1"
+                },
+                {
+                    "uses": "olafurpg/setup-scala@v2"
+                },
+                {
+                    "name": "Linting",
+                    "run": "csbt scalafixCheck githubActionsCheck"
+                }
             ]
-         },
-         "runs-on": "ubuntu-latest",
-         "steps": [
-            {
-               "uses": "actions/checkout@v1"
-            },
-            {
-               "uses": "olafurpg/setup-scala@v2"
-            },
-            {
-               "name": "Run tests",
-               "run": "csbt 'unit/testOnly -- tests.${{ matrix.test }}'"
-            }
-         ]
-      }
-   },
-   "name": "CI",
-   "on": [
-      "push",
-      "pull_request"
-   ]
+        },
+        "Docusaurus": {
+            "runs-on": "ubuntu-latest",
+            "steps": [
+                {
+                    "uses": "actions/checkout@v1"
+                },
+                {
+                    "uses": "olafurpg/setup-scala@v2"
+                },
+                {
+                    "name": "Website",
+                    "run": "csbt docs/docusaurusCreateSite"
+                }
+            ]
+        }
+    }
 }
-...
