--- ci/build.json
@@ -1,4 +1,0 @@
-{
-  "scalaVersions": ["2.13.0","2.13.1","2.12.8","2.12.9","2.12.10","2.12.7","2.11.12"],
-  "tests": ["{foldingrange,filewatch,hover,metals,references,statusbar,completions}","{ansi,workspacesymbol,bsp,builds,applyedit,compile,definition}","{http,formatting,documentsymbol,documenthighlight,codelens,tvp,mtags}"]
-}
